+++ app/src/main/java/eu/kanade/tachiyomi/ui/security/SecureActivityDelegate.kt
@@ -9,6 +9,7 @@ import androidx.biometric.BiometricManager
 import eu.kanade.tachiyomi.core.security.SecurityPreferences
 import eu.kanade.tachiyomi.data.preference.PreferenceValues
 import eu.kanade.tachiyomi.data.preference.PreferencesHelper
+import eu.kanade.tachiyomi.source.isIncognitoModeForSource
 import eu.kanade.tachiyomi.ui.main.SearchActivity
 import eu.kanade.tachiyomi.util.system.AuthenticatorUtil
 import uy.kohesive.injekt.injectLazy
@@ -21,8 +22,8 @@ object SecureActivityDelegate {
 
     var locked: Boolean = true
 
-    fun setSecure(activity: Activity?) {
-        val incognitoMode = preferences.incognitoMode().get()
+    fun setSecure(activity: Activity?, sourceId: Long? = null) {
+        val incognitoMode = isIncognitoModeForSource(sourceId, preferences)
         val enabled = when (preferences.secureScreen().get()) {
             PreferenceValues.SecureScreenMode.ALWAYS -> true
             PreferenceValues.SecureScreenMode.INCOGNITO -> incognitoMode